common/requirements.txt
Django==3.2.16
djangorestframework==3.14.0
django-environ==0.9.0